+++ packages/kysely/migrate/migrations/2025-02-03T08:48:00_refresh_tokens.ts
@@ -7,22 +7,31 @@ import { Database } from "../../src/db";
 export async function up(db: Kysely<Database>): Promise<void> {
   await db.schema
     .createTable("refresh_tokens")
-    .addColumn("tenant_id", "varchar(255)", (col) =>
-      col.references("tenants.id").onDelete("cascade").notNull(),
+    .addColumn("id", "varchar(21)", (col) => col.primaryKey())
+    .addColumn("client_id", "varchar(21)", (col) =>
+      col.references("applications.id").onDelete("cascade").notNull(),
     )
-    .addColumn("token", "varchar(255)", (col) => col.notNull())
-    .addColumn("session_id", "varchar(255)", (col) =>
-      col.references("sessions.session_id").onDelete("cascade").notNull(),
+    .addColumn("tenant_id", "varchar(255)")
+    // this is not a foreign key as the session could expire and be deleted
+    .addColumn("session_id", "varchar(21)", (col) => col.notNull())
+    .addColumn("user_id", "varchar(255)")
+    // same change here as on other tables - FK reference needed to users table
+    .addForeignKeyConstraint(
+      "user_id_constraint",
+      ["user_id", "tenant_id"],
+      "users",
+      ["user_id", "tenant_id"],
+      (cb) => cb.onDelete("cascade"),
     )
-    .addColumn("expires_at", "varchar(255)", (col) => col.notNull())
-    .addColumn("used_at", "varchar(255)")
-    .addColumn("scope", "varchar(512)", (col) => col.notNull())
-    .addColumn("audience", "varchar(512)", (col) => col.notNull())
-    .addColumn("revoked_at", "varchar(255)")
-    .addColumn("created_at", "varchar(255)", (col) =>
-      col.notNull().defaultTo(new Date().toISOString()),
-    )
-    .addPrimaryKeyConstraint("refresh_tokens_pkey", ["tenant_id", "token"])
+    .addColumn("created_at", "varchar(35)", (col) => col.notNull())
+    .addColumn("expires_at", "varchar(35)")
+    .addColumn("idle_expires_at", "varchar(35)")
+    .addColumn("last_exchanged_at", "varchar(35)")
+    // Contains a json blob with user agents.
+    .addColumn("device", "varchar(2048)", (col) => col.notNull())
+    // Contains a json blob with user agents.
+    .addColumn("resource_servers", "varchar(2048)", (col) => col.notNull())
+    .addColumn("rotating", "boolean", (col) => col.notNull())
     .execute();
 }