+++ packages/kysely/src/refreshTokens/get.ts
@@ -5,12 +5,12 @@ import { Database } from "../db";
 export function get(db: Kysely<Database>) {
   return async (
     tenant_id: string,
-    token: string,
+    id: string,
   ): Promise<RefreshToken | null> => {
     const refreshToken = await db
       .selectFrom("refresh_tokens")
       .where("refresh_tokens.tenant_id", "=", tenant_id)
-      .where("refresh_tokens.token", "=", token)
+      .where("refresh_tokens.id", "=", id)
       .selectAll()
       .executeTakeFirst();